Improve package publishing progress reporting
- increase status key expiration to 60mins
- report number of successful processing steps as a proportion of total
steps.

Note: total number of steps is a ghastly hard-coded value. This should
be reported from the dpp-fiscal plugin, but I don't think it's known
ahead of time. If changes are made to dpp-fiscal that change the number
of steps, this will also need to be changed to ensure accurate progress
reporting.